feat: added ReplicatedEntityExtension Gym (complete and tested)
notes: in this gym, a ReplicatedEntity is added as an EntityExtension which has an Ability that adds the CancelExtensionOwnerAbility tag. Because this is an EntityExtension, the ExtensionOwner gets the tags _and_ the Abilities. This cancels the ExtensionOwner's Ability that is continuously printint to screen. The EntityExtension is then destroyed, which in turn removes the tags and abilities on the ExtensionOwner and the ContinuousAbility can be re-activated if desired.