refactor: simplified the Base Interactable
notes: this is possible because of the ContextHandle which can be used
to add the Interactable and the Probe directly on the owning Actor.
Although the previous method is also perfectly acceptable (and a better
choice in certain circumstances e.g. when we want the Interactable to be
a separate Entity)